add file listings and subdir handling

+++ files.go
@@ -4,6 +4,8 @@ import (
 	"io/ioutil"
 	"net/http"
 	"os"
+	"path"
+	"strings"
 )
 
 func init() {
@@ -14,18 +16,56 @@ func init() {
 	http.Handle("/files/", http.StripPrefix("/files/", fs))
 }
 
-func Fetch(filename string) []byte {
-	// TODO: this isn't safe if the filename uses a nasty path!
-	data, _ := ioutil.ReadFile(Settings.FilesDir + "/" + filename)
-	return data
+func PathIsSafe(s string) bool {
+	// TODO: still a bit ad-hoc, with probably lots of holes in these checks
+	s = path.Clean(s)
+	return s != "" && !strings.HasPrefix(s, ".") && !strings.HasSuffix(s, "/")
 }
 
-func Store(filename string, body []byte) error {
-	// TODO: this isn't safe if the filename uses a nasty path!
-	fpath := Settings.FilesDir + "/" + filename
-	if len(body) > 0 {
-		return ioutil.WriteFile(fpath, body, 0666)
-	} else {
-		return os.Remove(fpath)
+func Fetch(filename string) (data []byte) {
+	if PathIsSafe(filename) {
+		data, _ = ioutil.ReadFile(Settings.FilesDir + "/" + filename)
 	}
+	return
+}
+
+func Store(filename string, body []byte) (err error) {
+	if PathIsSafe(filename) {
+		fpath := Settings.FilesDir + "/" + filename
+		if len(body) > 0 {
+			if err = os.MkdirAll(path.Dir(fpath), 0755); err == nil {
+				err = ioutil.WriteFile(fpath, body, 0666)
+			}
+		} else {
+			// iterate to automatically clean out empty parent dirs
+			for {
+				err = os.Remove(fpath)
+				if err != nil {
+					break
+				}
+				filename = path.Dir(filename)
+				if filename == "." {
+					break
+				}
+				fpath = Settings.FilesDir + "/" + filename
+				all, _ := ioutil.ReadDir(fpath)
+				if len(all) > 0 {
+					break
+				}
+			}
+		}
+	}
+	return
+}
+
+func FileList(dirname string, dir bool) (files []string) {
+	if dirname == "." || PathIsSafe(dirname) {
+		all, _ := ioutil.ReadDir(Settings.FilesDir + "/" + dirname)
+		for _, f := range all {
+			if f.IsDir() == dir {
+				files = append(files, f.Name())
+			}
+		}
+	}
+	return
 }
